+++ src/pages/RentalManagement/RentalManagementHeader.tsx
@@ -38,7 +38,8 @@ function RentalManagementHeader(props) {
     dispatch,
     showTransferEntityDialog,
     selectedType,
-    fetchRentalManagement
+    fetchRentalManagement,
+    gridApi
     // showCloneRentalManagementDialog
   } = props;
 
@@ -116,10 +117,13 @@ function RentalManagementHeader(props) {
       .then(({ data: { data } }) => {
         insertUpdate(objectStore.resource, 'serializedAsset', data);
       });
+    axiosInstance()
+      .get(`/field?resource=Product&view=true`)
+      .then(({ data: { data } }) => {
+        insertUpdate(objectStore.resource, 'Product', data);
+      });
+    if (gridApi) gridApi.deselectAll()
   };
-
-
-
 
   const handleRemoveoffline = async () => {
     await clearAll(objectStore.rentalManagement);
@@ -290,7 +294,7 @@ function RentalManagementHeader(props) {
                       } */}
                       {
                         <MenuItem
-                          disabled={!selectedRecords.length || selectedRecords.find((d) => d.canDelete === false)}
+                          disabled={!selectedRecords.length}
                           onClick={() => handleAddOffline()}
                         >
                           Add Offline