+++ src/pages/SerializedAsset/SerializedAssetDetailsPage.tsx
@@ -139,7 +139,7 @@ const SerializedAssetDetailsPage = () => {
   const columns = [
     { field: "reference", headerName: "Reference", show: true, cellRenderer: "nameRenderer" },
     { field: "type", headerName: "Type", show: true, disabled: true, cellRenderer: "commonRenderer" },
-    { field: "date", headerName: "Date & Time", show: true, disabled: true, cellRenderer: "dateTimeRenderer" },
+    { field: "date", headerName: "Date & Time", show: true, disabled: true, filter: false, cellRenderer: "dateTimeRenderer" },
     { field: "status", headerName: "Status", show: true, cellRenderer: "commonRenderer" },
     { field: "comments", headerName: "Comment", show: true, cellRenderer: "commonRenderer" },
     { field: "location", headerName: "Location", show: true, cellRenderer: "commonRenderer" },